Update `example_runScripts_SasView.py` and add `sas/sascalc` to FitBenchmarking
The file `example_runScripts_SasView.py` has been updated so that it will prompt a message if Bumps, sasmodels or sascalc is/are not installed.

The folder `sas/sascalc` is added as sascalc is not an independent package to be installed via `pip` yet. This folder should be removed once there is an installation method for sascalc.