Issue 693: Add sudo to dockerfile
Azure pipelines adds a new user with sudo permissions to run tests
in a container. This means that the user has no root access on standard
machines as these do not include sudo (?!).

Add sudo here so that pipelines can install python packages to the venv
that we create with the root (default) account.