Issue #682: Checkout branch to submit coverage
Coveralls pulls information from the git repo when the report is
submitted. This has been developed a lot for travis support, but is
lagging for other CI/CD. The best solution at present is to checkout the
branch before submitting the report.

I imagine this could have issues if branches are editted while tests are
run, however in that case I'd expect them to be rerun anyway!?