Issue #682: Move to GitHub Actions; Travis stopped supporting open source (#692)
* Issue #682: Move to Azure; Travis stopped supporting open source

Azure creates a new user to run tests in.
Using sudo causes issues with paths.
So make we chown cache folders to the current user via sudo before
execution, and the tests work!!!

* Issue #682: Link coverage report to azure

* Issue 693: Add sudo to dockerfile

Azure pipelines adds a new user with sudo permissions to run tests
in a container. This means that the user has no root access on standard
machines as these do not include sudo (?!).

Add sudo here so that pipelines can install python packages to the venv
that we create with the root (default) account.

* Issue 682: Update coverage

* Issue #682: Add token to connect azure to coveralls. Token stored by azure

* Issue #682: Checkout branch to submit coverage

Coveralls pulls information from the git repo when the report is
submitted. This has been developed a lot for travis support, but is
lagging for other CI/CD. The best solution at present is to checkout the
branch before submitting the report.

I imagine this could have issues if branches are editted while tests are
run, however in that case I'd expect them to be rerun anyway!?

* Point travis at new tests location

* Issue #682: Remove travis from repo

* Issue #682: Add azure scheduled tests

* Issue #682: Impersonate CircleCI for coveralls

* insert missing "$"

* testing coveralls

* testing coveralls

* testing coveralls

* testing coveralls

* testing coveralls

* Try github actions...

* Trigger rerun

* Allow manual starting

* Rename workflow

* Debugging actions

* Debugging actions

* Debugging actions

* Debugging actions

* Debugging actions

* Debugging actions

* Debugging actions

* Debugging actions

* Connect actions to coveralls

* Connect actions to coveralls

* Remove pipelines in favour of GitHub Actions

* Fix typo in test name

* Add linting to github actions

Co-authored-by: Tyrone Rees <tyrone.rees@stfc.ac.uk>